add new RecordManager delete html doc(will be added in website)
Former-commit-id: 98e4dfefaaa04154b718c0c6dc31baa4d04d33aa [formerly fa63020a7db5869059d23eab3e39a30b2b777138] [formerly 1ce0f391b89852acc71555d39211b740a3bd2ac7] [formerly 5adb8d98bcc4c4b576e97a4293e09dfa90242f4f [formerly 27d07d5ed5ae8843dd119e193bc15136820c9fc4]] [formerly d6b0ff1a179132a1d9ed8fe621e825e89791fd4d [formerly 004676c55be0223a55a4983c107a890643c97587] [formerly 5a3b158341574d63b4b31abbb7274e1ea92feeef [formerly 1fbcf081cffcf7cc5772021e53e87c9310f59d43]]] [formerly 95d1be615f80c4ec5b823828098e895a9c8daaf5]
Former-commit-id: 25fb942242c78830cd3987464882c64d5cf2086a
Former-commit-id: 8b547b4dc83a52641d6b6fb9a0c0b64bc6c74885 [formerly f886fea6f95a28de87cff5567eb4055be961fbb2] [formerly c1e94f8ed16df60313d3ef0c6a10cf70342e0334 [formerly 2a3b75299a9b5abe56426dee26ce00f41b34908e]]
Former-commit-id: 2b98e8373d1d8f5d6c8fb208252fe34dca451aa4 [formerly 3109e51b14222b1daafbe4f6d2d0d85dec8a4267]
Former-commit-id: 852225ad6c5c69b80bd8c4ba4f8c219f73663089
Former-commit-id: 0f2aa9ff06f4add6dc67b5d8e37b6bf8ca2d81f9

+++ Frame/SDK/Base/AFTime.hpp
@@ -550,7 +550,7 @@ public:
     {
         AFTime xTime((int)Get(YEAR), (int)Get(MONTH), 1, 0, 0, 0, 0);
         int nFirstDayOfWeek = xTime.GetDayOfWeek();
-        int nDaySpan = Get(DAY) - 1;
+        int64_t nDaySpan = Get(DAY) - 1;
         return (nFirstDayOfWeek + nDaySpan) / 7;
     }
